requirements.txt
anthropic
scikit-learn
numpy
torch
python-dotenv
langchain
langchain-core
langchain-astradb
langchain-anthropic
langchain_community
langchain_graph_retriever
langchain-huggingface
langchain_ibm
langchain-text-splitters
graph_retriever
ibm-watsonx-ai
keybert
pyvis
networkx
